sui-api/requirements.txt
tabulate==0.9.0
pandas==2.2.3
requests==2.32.3
python-dotenv==1.0.1
fastapi[standard]
loguru==0.7.2
groq==0.18.0
atoma-sdk==0.1.1